NU-2436 move selectors to testing.ts
