+++ designer/client/src/components/modals/AdhocTesting/useAdhocTestingAction.tsx
@@ -2,7 +2,8 @@ import { head } from "lodash";
 import { useCallback, useMemo } from "react";
 
 import { testProcessWithParameters } from "../../../actions/nk/displayTestResults";
-import { getProcessingType, getTestData, getTestParameters } from "../../../reducers/selectors/graph";
+import { getProcessingType, getTestParameters } from "../../../reducers/selectors/graph";
+import { getTestData } from "../../../reducers/selectors/testing";
 import { useAppDispatch, useAppSelector } from "../../../store/storeHelpers";
 import type { UIParameter } from "../../../types/definition";
 import { getFindAvailableVariables } from "../../graph/node-modal/NodeDetailsContent/selectors";
