+++ designer/client/src/components/modals/TestingDataRecords/useDataRecordsValidation.ts
@@ -1,14 +1,14 @@
 import React, { useState } from "react";
 
-import { getTestingDataRecords } from "../../../reducers/selectors/graph";
 import { getMaxTestingRecords } from "../../../reducers/selectors/settings";
+import { getInputDataRecords } from "../../../reducers/selectors/testCases";
 import { useAppSelector } from "../../../store/storeHelpers";
 
 type RecordError = { type: "TEST_DATA_LIMIT_EXCEEDED" };
 
 export const useDataRecordsValidation = () => {
     const maxTestingRecords = useAppSelector(getMaxTestingRecords);
-    const testingDataRecords = useAppSelector(getTestingDataRecords);
+    const testingDataRecords = useAppSelector(getInputDataRecords);
 
     const [recordsErrors, setRecordsErrors] = useState<RecordError[]>([]);
