+++ designer/client/src/reducers/selectors/graph.ts
@@ -2,7 +2,6 @@ import { isEmpty, isEqual } from "lodash";
 import { createSelector } from "reselect";
 
 import type { TestFormParameters } from "../../common/TestResultUtils";
-import type { ExpressionObj } from "../../components/graph/node-modal/editors/expression/types";
 import { hasNodeIdPlaceholder } from "../../components/graph/node-modal/nodeIdFieldHelpers";
 import ProcessStateUtils from "../../components/Process/ProcessStateUtils";
 import type { Scenario } from "../../components/Process/types";
@@ -10,7 +9,6 @@ import { isStatusRunning } from "../../components/Process/types";
 import { ScenarioGraphSourceType } from "../../http/HttpService/types";
 import type { ProcessCounts } from "../../http/resultsWithCountsDto";
 import type { ScenarioGraph } from "../../types/scenarioGraph";
-import type { TestData, TestingState } from "../graph/types";
 import type { RootState } from "../index";
 import { getHistoryPast } from "./getHistory";
 import { areLabelsUpdated, isGraphUpdated } from "./helpers";
@@ -27,8 +25,6 @@ export const getScenarioGraph = createSelector(getGraph, (g) => g.scenario.scena
         resultEqualityCheck: isEqual,
     },
 });
-
-export const getTesting = createSelector(getGraph, (g) => g.testing[g.scenario.name] || ({} as TestingState));
 
 export const getNodes = createSelector(getScenarioGraph, (g) => g.nodes);
 
@@ -76,35 +72,9 @@ export const isArchivePossible = createSelector(
     (state, isFragment) => isFragment || ProcessStateUtils.canArchive(state),
 );
 export const getTestCapabilities = createSelector(getGraph, (g) => g.testCapabilities);
-export const getTestingDataRecords = createSelector(getTesting, (g) => g.testingDataRecords || []);
-
-const getSourceId = (_: unknown, sourceId: string) => sourceId;
-export const getTestingDataRecordsForSingleSource = createSelector(
-    [getTestingDataRecords, getSourceId],
-    (testingDataRecords, sourceId: string) => testingDataRecords.filter((r) => r.sourceId === sourceId),
-);
-
-const getNodeId = (_: unknown, sourceId: string) => sourceId;
-
-export const hasTestingDataRecordsDefined = createSelector(getTestingDataRecords, (testingDataRecords) => testingDataRecords.length > 0);
-
 export const getTestParameters = createSelector(getGraph, (g) => g.testFormParameters || ([] as TestFormParameters[]));
-export const getTestResults = createSelector(getTesting, (g) => g.testResults);
-export const getTestAssertionResults = createSelector(getTesting, (g) => g.assertionsResults || {});
-export const getTestAssertionResultsForNode = createSelector(
-    getTestAssertionResults,
-    getNodeId,
-    (assertionsResults, nodeId) => assertionsResults[nodeId],
-);
-export const getTestResultsLoading = createSelector(getTesting, (g) => g.testResultsLoading);
-export const getTestData = createSelector(getTesting, (g) => g.testData || ({} as TestData));
 export const getProcessCountsRefresh = createSelector(getGraph, (g) => g.processCountsRefresh || null);
 export const getProcessCounts = createSelector(getGraph, (g): ProcessCounts => g.processCounts || ({} as ProcessCounts));
-export const getIsTestingMode = createSelector(
-    getTestResults,
-    getProcessCounts,
-    (results, counts) => !isEmpty(results) || !isEmpty(counts),
-);
 
 export const getVersions = createSelector(getScenario, (details) => details?.history || []);
 export const hasOneVersion = createSelector(getVersions, (h) => h.length <= 1);
